% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_dc.R
\name{create_dc}
\alias{create_dc}
\title{Find out whether a zip code has destination charger in 10 mile radius}
\usage{
create_dc()
}
\value{

}
\description{
Find out whether a zip code has destination charger in 10 mile radius
}
