% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_evtrip_scenarios.R
\name{create_departure_df}
\alias{create_departure_df}
\title{Create a dataframe that predicts the number of trips departing from origin to destination}
\usage{
create_departure_df(od, od_sp, config)
}
\arguments{
\item{od}{A dataframe containing the results of the gravity model}

\item{od_sp}{Dataframe containing OD and shortest_path_length}

\item{config}{constants}
}
\value{
A dataframe with column origin, destination, and
number of trips between the origin and destination
}
\description{
Create a dataframe that predicts the number of trips departing from origin to destination
}
