% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_evtrip_scenarios.R
\name{get_Gas_Price}
\alias{get_Gas_Price}
\title{This function returns the gas price for the zip code from the gas price dataframe}
\usage{
get_Gas_Price(gas_prices, origin_zip)
}
\arguments{
\item{gas_prices}{The dataframe containing gas price for each zip code}

\item{origin_zip}{The zip for which gas price is to be known}
}
\value{
The gas price for the relevant zip code
}
\description{
This function returns the gas price for the zip code from the gas price dataframe
}
