% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_wa_gas_prices.R
\name{get_gas_prices}
\alias{get_gas_prices}
\title{get gas prices from gas buddy API}
\usage{
get_gas_prices(zip)
}
\value{
TRUE if update was successful
}
\description{
Run tripgen::update_wa_bevs() to update the BEVs in the state.
}
