% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_evtrip_scenarios.R
\name{get_max_spacing}
\alias{get_max_spacing}
\title{Get max spacing between charging stations}
\usage{
get_max_spacing(main_con, a_id, origin, dest, connector_code)
}
\arguments{
\item{main_con}{The database connection}

\item{a_id}{The analysis_id}

\item{origin}{The origin zip code}

\item{connector_code}{The connector_code for the EV - ChaDEMO or COMBO}

\item{destination}{The destination zip code}
}
\value{
max_spacing The maximum spacing (in miles) between charging stations along the shortest path
}
\description{
Get max spacing between charging stations
}
