% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_evtrip_scenarios.R
\name{get_tripEVs_from_sourceEVs}
\alias{get_tripEVs_from_sourceEVs}
\title{Get the trips EVs from the relevant EVs belonging to a source zip code}
\usage{
get_tripEVs_from_sourceEVs(trips_source_i, source_EVs)
}
\arguments{
\item{trips_source_i}{Number of trips from a source zip code}

\item{source_EVs}{the dataframe of EVs from the zip code}
}
\value{
The selected trip EVs from the set of EVs at source
}
\description{
Get the trips EVs from the relevant EVs belonging to a source zip code
}
