% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_evtrip_scenarios.R
\name{rnd_date_time}
\alias{rnd_date_time}
\title{Generate random date times between two date-times, for a timezone
Adapted from : https://stackoverflow.com/a/14721124/1328232}
\usage{
rnd_date_time(
  N,
  st = "2019/07/01",
  et = "2019/07/02",
  tz = "America/Los_Angeles",
  config
)
}
\arguments{
\item{N}{Number of datetime values to generate}

\item{st}{Start date in the format "yyyy/mm/dd"}

\item{et}{End date in the format "yyyy/mm/dd"}

\item{tz}{Timezone for the dates}

\item{config}{constants}
}
\value{
The randomly generated date time.
}
\description{
Generate random date times between two date-times, for a timezone
Adapted from : https://stackoverflow.com/a/14721124/1328232
}
