% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_bevs.R, R/update_wa_gas_prices.R
\name{sendEmail}
\alias{sendEmail}
\title{Email a user a report is ready}
\usage{
sendEmail(email = "cp84@uw.edu", mail_message = "gas prices have NAs")

sendEmail(email = "cp84@uw.edu", mail_message = "gas prices have NAs")
}
\arguments{
\item{email}{Email to send to}

\item{mail_message}{Any extra info}
}
\value{
TRUE if successful email sent

TRUE if successful email sent
}
\description{
Requires an account at Mailgun: https://mailgun.com
Pre-verification can only send to a whitelist of emails you configure

Requires an account at Mailgun: https://mailgun.com
Pre-verification can only send to a whitelist of emails you configure
}
