% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_evtrip_scenarios.R
\name{trip_gen}
\alias{trip_gen}
\title{Create the EV fleet for ABM simulation based on SDCM VCDM}
\usage{
trip_gen(num_days = 1, config, a_id = 15)
}
\arguments{
\item{num_days}{Number of days to generate the EV fleet for.}

\item{config}{constants}

\item{a_id}{analysis_id}
}
\value{
Day EVs
}
\description{
This function `trip_gen()` creates a fleet of EVs for EVI-ABM simulation
using the published models or vehicle choice using static choice discrete models.
}
