% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_cd.R
\name{update_cd}
\alias{update_cd}
\title{Update the charging distances}
\usage{
update_cd(config, a_id)
}
\arguments{
\item{config}{constants}

\item{a_id}{analysis_id}
}
\value{

}
\description{
This function `update_cd()` updates the charging distances given the new set of charging stations.
}
