% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_dc.R
\name{update_dc}
\alias{update_dc}
\title{Find if there is a destination charger}
\usage{
update_dc(a_id = 1)
}
\arguments{
\item{a_id}{analysis_id}
}
\value{

}
\description{
Find if there is a destination charger
}
