% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_evses.R
\name{update_evses}
\alias{update_evses}
\title{Update the EVSEs in the database from the AFDC portal}
\usage{
update_evses()
}
\value{
TRUE if update was successful
}
\description{
Run tripgen::update_evses() to update the EVSEs table. Will overwrite. Needs AFDC API key.
}
