% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_bevs.R
\name{update_wa_bevs}
\alias{update_wa_bevs}
\title{Update the BEVs in the database by fetching the latest from WADOL and overwriting}
\usage{
update_wa_bevs()
}
\value{
TRUE if update was successful
}
\description{
Run tripgen::update_wa_bevs() to update the BEVs in the state.
}
