% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_wa_gas_prices.R
\name{update_wa_gas_prices}
\alias{update_wa_gas_prices}
\title{Update the gas prices in the database by fetching the latest from gas buddy and overwriting}
\usage{
update_wa_gas_prices()
}
\value{
TRUE if update was successful
}
\description{
Run tripgen::update_wa_bevs() to update the BEVs in the state.
}
