% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_evtrip_scenarios.R
\name{vcdm_scdm4}
\alias{vcdm_scdm4}
\title{Predicts the probability of choice of ICEVs vs Rental vehicle vs an EV for a trip
Taken from Yan's paper}
\usage{
vcdm_scdm4(ev_range, trip_row, config)
}
\arguments{
\item{ev_range}{- range of the EV chosen for the trip}

\item{trip_row}{Dataframe that gives trip relates parameters -
dist - trip distance
dest_charger_L2 - whether trip destination has a Level-2 Charger
dest_charger - whether destination charger has a DCFC
max_spacing - Maximum gap between DCFC chargers along the shortest path for the trip
gas_price - Price of gas at the origin zip code of the trip}

\item{config}{constants}
}
\value{
The probability of travel by EV
}
\description{
Predicts the probability of choice of ICEVs vs Rental vehicle vs an EV for a trip
Taken from Yan's paper
}
